fix logout cant load website error
